feat: 로또 번호 매칭 결과를 담기위한 클래스 구현

+++ src/main/java/lotto/util/ProfitCalculator.java
@@ -1,10 +1,8 @@
 package lotto.util;
 
-import java.text.DecimalFormat;
-
 public class ProfitCalculator {
 
-    public static double calculateProfitPercentage(double purchaseAmount, double winningAmount) {
+    public static double calculateProfitPercentage(double purchaseAmount, long winningAmount) {
         double profitPercentage = (winningAmount / purchaseAmount) * 100.0;
         return roundToTwoDecimalPlaces(profitPercentage);
     }
